[fix] #111 회원가입 완료 화면에서 뒤로가기 비활성화
