+++ Projects/Feature/FeatureLogin/Sources/LoginRoot/LoginRootView.swift
@@ -1,19 +1,17 @@
 //
-//  SignUpNavigationStackView.swift
+//  LoginRootView.swift
 //  Feature
 //
-//  Created by 김도형 on 7/7/24.
+//  Created by 김도형 on 9/7/24.
 
-import ComposableArchitecture
 import SwiftUI
 
-import DSKit
+import ComposableArchitecture
 
-@ViewAction(for: LoginRootFeature.self)
 public struct LoginRootView: View {
     /// - Properties
-    @Perception.Bindable
-    public var store: StoreOf<LoginRootFeature>
+    public let store: StoreOf<LoginRootFeature>
+    
     /// - Initializer
     public init(store: StoreOf<LoginRootFeature>) {
         self.store = store
@@ -23,35 +21,14 @@ public struct LoginRootView: View {
 public extension LoginRootView {
     var body: some View {
         WithPerceptionTracking {
-            NavigationStack(path: $store.scope(state: \.path, action: \.path)) {
-                WithPerceptionTracking {
-                    VStack(spacing: 8) {
-                        logo
-                        
-                        Spacer()
-                        
-                        appleLoginButton
-                            .pokitMaxWidth()
-                        
-                        googleLoginButton
-                            .pokitMaxWidth()
+            Group {
+                switch store.state {
+                case .login:
+                    if let store = store.scope(state: \.login, action: \.login) {
+                        LoginView(store: store)
                     }
-                    .padding(.horizontal, 20)
-                    .padding(.bottom, 36)
-                    .background(.pokit(.bg(.base)))
-                    .ignoresSafeArea(edges: .bottom)
-                    .navigationBarBackButtonHidden()
-                }
-            } destination: { path in
-                WithPerceptionTracking {
-                    switch path.case {
-                    case .agreeToTerms(let store):
-                        AgreeToTermsView(store: store)
-                    case .registerNickname(let store):
-                        RegisterNicknameView(store: store)
-                    case .selecteField(let store):
-                        SelectFieldView(store: store)
-                    case .signUpDone(let store):
+                case .signUpDone:
+                    if let store = store.scope(state: \.signUpDone, action: \.signUpDone) {
                         SignUpDoneView(store: store)
                     }
                 }
@@ -60,114 +37,17 @@ public extension LoginRootView {
     }
 }
 //MARK: - Configure View
-extension LoginRootView {
-    private var logo: some View {
-        HStack {
-            Spacer()
-            
-            VStack(spacing: 24) {
-                Image(.logo(.pokit))
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(height: 72)
-                    .foregroundStyle(.pokit(.icon(.brand)))
-                
-                Text("다양한 링크들을 한 곳에")
-                    .pokitFont(.b1(.b))
-                    .foregroundStyle(.pokit(.text(.secondary)))
-            }
-            
-            Spacer()
-        }
-        .padding(.top, 254)
-    }
+private extension LoginRootView {
     
-    private var appleLoginButton: some View {
-        Button {
-            send(.appleLoginButtonTapped)
-        } label: {
-            appleLoginButtonLabel
-        }
-    }
-    
-    private var appleLoginButtonLabel: some View {
-        VStack {
-            Spacer()
-            
-            HStack(spacing: 12) {
-                Spacer()
-                
-                Image(systemName: "apple.logo")
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(width: 24, height: 24)
-                    .foregroundColor(.pokit(.icon(.inverseWh)))
-                
-                Text("Apple로 계속하기")
-                    .pokitFont(.l1(.r))
-                    .foregroundStyle(.pokit(.text(.inverseWh)))
-                
-                Spacer()
-            }
-            
-            Spacer()
-        }
-        .background {
-            RoundedRectangle(cornerRadius: 8, style: .continuous)
-                .fill(.black)
-                .overlay {
-                    RoundedRectangle(cornerRadius: 8, style: .continuous)
-                        .stroke(.pokit(.border(.secondary)), lineWidth: 1)
-                }
-        }
-        .frame(height: 50)
-    }
-    
-    private var googleLoginButton: some View {
-        Button {
-            send(.googleLoginButtonTapped)
-        } label: {
-            googleLoginButtonLabel
-        }
-    }
-    
-    private var googleLoginButtonLabel: some View {
-        VStack {
-            Spacer()
-            
-            HStack(spacing: 12) {
-                Spacer()
-                
-                Image(.icon(.google))
-                    .resizable()
-                    .frame(width: 24, height: 24)
-                
-                Text("Google로 계속하기")
-                    .pokitFont(.l1(.r))
-                    .foregroundStyle(.pokit(.text(.primary)))
-                
-                Spacer()
-            }
-            
-            Spacer()
-        }
-        .background {
-            RoundedRectangle(cornerRadius: 8, style: .continuous)
-                .fill(.pokit(.bg(.base)))
-                .overlay {
-                    RoundedRectangle(cornerRadius: 8, style: .continuous)
-                        .stroke(.pokit(.border(.secondary)), lineWidth: 1)
-                }
-        }
-        .frame(height: 50)
-    }
 }
 //MARK: - Preview
 #Preview {
     LoginRootView(
         store: Store(
-            initialState: .init(),
+            initialState: .login(.init()),
             reducer: { LoginRootFeature() }
         )
     )
 }
+
+
